+++ src/command_modules/azure-cli-monitor/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+
+unreleased
+++++++++++
+* Add action-group commands
+
 0.0.10 (2017-09-22)
 +++++++++++++++++++
 * minor fixes